requirements.txt
azure-devops==7.1.0b4
numpy==2.2.0
pandas==2.2.3
atlassian-python-api==3.41.19
requests==2.32.3
PyYAML==6.0.2
Jinja2==3.1.5
MarkupSafe==3.0.2